Fix: use same matrix type (i64) across all languages

+++ comptime/comptime.rs
@@ -3,8 +3,8 @@ use std::time::{Instant};
 #[derive(Debug)]
 struct Matrix<T, const M: usize, const N: usize> ([[T; N]; M]);
 
-const fn cross<const M: usize, const N: usize, const K: usize> (lhs: Matrix<isize, M, K>, rhs: Matrix<isize, K, N>) -> Matrix<isize, M, N> {
-    let mut new: Matrix<isize,M,N> = Matrix([[0isize; N]; M]);
+const fn cross<const M: usize, const N: usize, const K: usize> (lhs: Matrix<i64, M, K>, rhs: Matrix<i64, K, N>) -> Matrix<i64, M, N> {
+    let mut new: Matrix<i64,M,N> = Matrix([[0i64; N]; M]);
 
     let mut row = 0usize;
     while row < M {
@@ -12,7 +12,7 @@ const fn cross<const M: usize, const N: usize, const K: usize> (lhs: Matrix<isiz
         let mut col = 0usize;
         while col < N {
         //for col in 0..N {
-            let mut sum = 0isize;
+            let mut sum = 0i64;
             let mut j   = 0usize;
             while j < K {
             //for j in 0..K {
@@ -31,12 +31,12 @@ const fn cross<const M: usize, const N: usize, const K: usize> (lhs: Matrix<isiz
 }
 
 fn main() {
-    const left: Matrix<isize,50,20> = Matrix([[5; 20]; 50]);
-    const right: Matrix<isize,20,15> = Matrix([[2; 15]; 20]);
+    const left: Matrix<i64,50,20> = Matrix([[5; 20]; 50]);
+    const right: Matrix<i64,20,15> = Matrix([[2; 15]; 20]);
 
     let now = Instant::now();
 
-    const result: Matrix<isize, 50_usize, 15_usize> = cross(left, right);
+    const result: Matrix<i64, 50_usize, 15_usize> = cross(left, right);
 
     let elapsed = now.elapsed().as_nanos();
