+++ runtime/runtime-get-if-let.rs
@@ -3,13 +3,13 @@ use std::time::{Instant};
 #[derive(Debug)]
 struct Matrix<T, const M: usize, const N: usize> ([[T; N]; M]);
 
-fn cross<const M: usize, const N: usize, const K: usize> (lhs: Matrix<isize, M, K>, rhs: Matrix<isize, K, N>) -> Matrix<isize, M, N> {
-    let mut new: Matrix<isize,M,N> = Matrix([[0isize; N]; M]);
+fn cross<const M: usize, const N: usize, const K: usize> (lhs: Matrix<i64, M, K>, rhs: Matrix<i64, K, N>) -> Matrix<i64, M, N> {
+    let mut new: Matrix<i64,M,N> = Matrix([[0i64; N]; M]);
 
     for row in 0..M {
         if let Some(new_row) = new.0.get_mut(row) {
             for col in 0..N {
-                let mut sum = 0isize;
+                let mut sum = 0i64;
 
                 if let Some(cell) = new_row.get_mut(col) {
 
@@ -26,12 +26,12 @@ fn cross<const M: usize, const N: usize, const K: usize> (lhs: Matrix<isize, M, 
 }
 
 fn main() {
-    let left: Matrix<isize,50,20> = Matrix([[5; 20]; 50]);
-    let right: Matrix<isize,20,15> = Matrix([[2; 15]; 20]);
+    let left: Matrix<i64,50,20> = Matrix([[5; 20]; 50]);
+    let right: Matrix<i64,20,15> = Matrix([[2; 15]; 20]);
 
     let now = Instant::now();
 
-    let result: Matrix<isize, 50_usize, 15_usize> = cross(left, right);
+    let result: Matrix<i64, 50_usize, 15_usize> = cross(left, right);
 
     let elapsed = now.elapsed().as_nanos();
